- We have changed the use case to build the datasets. Now, those columns where all the values are zero will be removed from the final generated CSV file.
- We have added a new method for dataset balancing based on
synthetic data generation, the Gaussian_Copula, that takes into
account the correlation among the variables of the original
dataset.

- We have included a new step in the pipeline, the imputation of
the missing values. At the moment, it has the option to use the
kNN algorithm to do the task and another option to not perform
the imputation.

- We have changed the implementation of the normalization to fill
with zeros those columns that have at least one NaN.

- We have modified one of the scripts that were implemented to
automatize the launching of experiments' sets in batch.

- Now, both train and test will store the ROC curve on its
corresponding PDF file.