- We have removed some files that are not necessary anymore.
- We have separated the use case to build the datasets in 3
different scenarios:

    - Scenario I: only clinical data.

    - Scenario II: only imaging data.

    - Scenario III: clinical data + imaging data.

- Regarding the previous point, we have implemented 2 different
methods for each scenario. The first one will build a full version
of the dataset (with all the columns) and the second one will
provide a simplified version (with only the needed column for the
training process).

- We have modified the test script to obtain the metrics with
different operation points.

- We have added a new script that does the following: given a
directory with a set of CSV files with the results of an
experiments' batch, it generates a single CSV file that compares
all of them more easily.

- We include a script to preprocess the original dataset. The
preprocessed version of the original dataset allows to generate
the built datasets in a more straightforward way.