We have made the following changes:
- The scripts of training and test were changed to allow the option
to pass "all" as a possible argument for the number of features,
that was previously just an integer. This change was also applied
to "obtain_features_ranking.py".

- We have improved the implementation of the script
"obtain_metrics_comparison.py" to include the mean and the standard
deviation of each metric.

- An automatically generated "requirements.txt" was added to the
repository.

- We have changed the code that plots the ranking of the features,
to only include the clinical ones. Previously, all the features
were shown, making the generated plot impossible to be read. Now,
it includes the names of the features and their global position.

- We have also performed some changes to the scripts included within
the directory "execution_scripts", that automatize the processes
of training, specially useful when we are running many sequential
training processes in just one call.